requirements.txt
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
oauthlib==3.2.2
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.8.0
six==1.17.0
sniffio==1.3.1
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
python-dotenv
